Renamed endian.h to neocd_endian.h to avoid conflicts under Linux. Updated libretro-common submodule
