+++ src/3rdparty/z80/z80.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif
 
-#include "../../endian.h"
+#include "../../neocd_endian.h"
 
 #include <stdint.h>
