+++ src/neocd_endian.h
